rewrite the "history inspection"; fixes #175, fixes #181
instead of having a lot of text which is presented as demo,
we rather introduce 4 commands that we need and let participants
clone an (R) repository and in there they are asked to:
- find the line which contains a string (an error message)
- find out when that line was last changed
- find the corresponding commit
- inspect that commit with git show
- bring the repo to a state at time of that commit

I think/hope that this will be more fun for the participants
than seeing a lot of text and having those commands presented
by the instructor. I encourage to do this exercise in groups of two.